Rework typing preference screen and integrate spelling
