+++ app/src/main/kotlin/dev/patrickgold/florisboard/app/settings/typing/TypingScreen.kt
@@ -18,9 +18,12 @@ package dev.patrickgold.florisboard.app.settings.typing
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import dev.patrickgold.florisboard.R
+import dev.patrickgold.florisboard.ime.nlp.SpellingLanguageMode
 import dev.patrickgold.florisboard.ime.smartbar.CandidatesDisplayMode
 import dev.patrickgold.florisboard.lib.android.AndroidVersion
 import dev.patrickgold.florisboard.lib.compose.FlorisErrorCard
@@ -39,31 +42,31 @@ fun TypingScreen() = FlorisScreen {
     previewFieldVisible = true
 
     content {
+        // This card is temporary and is therefore not using a string resource
+        FlorisErrorCard(
+            modifier = Modifier.padding(8.dp),
+            text = """
+                Suggestions (except system autofill) and spell checking are not available in this alpha release. Most
+                preferences for these are placeholders at the moment.
+            """.trimIndent().replace('\n', ' '),
+        )
+
         PreferenceGroup(title = stringRes(R.string.pref__suggestion__title)) {
-            SwitchPreference(
-                prefs.suggestion.api30InlineSuggestionsEnabled,
-                title = stringRes(R.string.pref__suggestion__api30_inline_suggestions_enabled__label),
-                summary = stringRes(R.string.pref__suggestion__api30_inline_suggestions_enabled__summary),
-                visibleIf = { AndroidVersion.ATLEAST_API30_R },
-            )
-            // This card is temporary and is therefore not using a string resource
-            FlorisErrorCard(
-                modifier = Modifier.padding(8.dp),
-                text = if (AndroidVersion.ATLEAST_API30_R) {
-                    "Suggestions (except autofill) are not available in this release"
-                } else {
-                    "Suggestions are not available in this release"
-                },
-            )
             SwitchPreference(
                 prefs.suggestion.enabled,
                 title = stringRes(R.string.pref__suggestion__enabled__label),
-                //summary = stringRes(R.string.pref__suggestion__enabled__summary),
+                summary = stringRes(R.string.pref__suggestion__enabled__summary),
             )
             ListPreference(
                 prefs.suggestion.displayMode,
                 title = stringRes(R.string.pref__suggestion__display_mode__label),
                 entries = CandidatesDisplayMode.listEntries(),
+                enabledIf = { prefs.suggestion.enabled isEqualTo true },
+            )
+            SwitchPreference(
+                prefs.suggestion.blockPossiblyOffensive,
+                title = stringRes(R.string.pref__suggestion__block_possibly_offensive__label),
+                summary = stringRes(R.string.pref__suggestion__block_possibly_offensive__summary),
                 enabledIf = { prefs.suggestion.enabled isEqualTo true },
             )
             SwitchPreference(
@@ -75,13 +78,18 @@ fun TypingScreen() = FlorisScreen {
             DialogSliderPreference(
                 prefs.suggestion.clipboardContentTimeout,
                 title = stringRes(R.string.pref__suggestion__clipboard_content_timeout__label),
-                valueLabel = { stringRes(R.string.unit__seconds__symbol, "v" to it) },
+                valueLabel = { stringRes(R.string.pref__suggestion__clipboard_content_timeout__summary, "v" to it) },
                 min = 30,
                 max = 300,
                 stepIncrement = 5,
-                enabledIf = {
-                    (prefs.suggestion.enabled isEqualTo true) && (prefs.suggestion.clipboardContentEnabled isEqualTo true)
-                }
+                enabledIf = { prefs.suggestion.enabled isEqualTo true },
+                visibleIf = { prefs.suggestion.clipboardContentEnabled isEqualTo true },
+            )
+            SwitchPreference(
+                prefs.suggestion.api30InlineSuggestionsEnabled,
+                title = stringRes(R.string.pref__suggestion__api30_inline_suggestions_enabled__label),
+                summary = stringRes(R.string.pref__suggestion__api30_inline_suggestions_enabled__summary),
+                visibleIf = { AndroidVersion.ATLEAST_API30_R },
             )
         }
 
@@ -100,6 +108,34 @@ fun TypingScreen() = FlorisScreen {
                 prefs.correction.doubleSpacePeriod,
                 title = stringRes(R.string.pref__correction__double_space_period__label),
                 summary = stringRes(R.string.pref__correction__double_space_period__summary),
+            )
+        }
+
+        PreferenceGroup(title = stringRes(R.string.pref__spelling__title)) {
+            val florisSpellCheckerEnabled = remember { mutableStateOf(false) }
+            SpellCheckerServiceSelector(florisSpellCheckerEnabled)
+            ListPreference(
+                prefs.spelling.languageMode,
+                iconId = R.drawable.ic_language,
+                title = stringRes(R.string.pref__spelling__language_mode__label),
+                entries = SpellingLanguageMode.listEntries(),
+                enabledIf = { florisSpellCheckerEnabled.value },
+            )
+            SwitchPreference(
+                prefs.spelling.useContacts,
+                iconId = R.drawable.ic_contacts,
+                title = stringRes(R.string.pref__spelling__use_contacts__label),
+                summary = stringRes(R.string.pref__spelling__use_contacts__summary),
+                enabledIf = { florisSpellCheckerEnabled.value },
+                visibleIf = { false }, // For now
+            )
+            SwitchPreference(
+                prefs.spelling.useUdmEntries,
+                iconId = R.drawable.ic_library_books,
+                title = stringRes(R.string.pref__spelling__use_udm_entries__label),
+                summary = stringRes(R.string.pref__spelling__use_udm_entries__summary),
+                enabledIf = { florisSpellCheckerEnabled.value },
+                visibleIf = { false }, // For now
             )
         }
     }
