+++ app/src/main/res/values-de/strings.xml
@@ -345,13 +345,8 @@
     <string name="pref__advanced__force_private_mode__label" comment="Label of Force private mode preference in Advanced">Privaten Modus erzwingen</string>
     <string name="pref__advanced__force_private_mode__summary" comment="Summary of Force private mode preference in Advanced">Deaktiviert Funktionen die mit Eingabedaten arbeiten</string>
     <!-- Spelling UI strings -->
-    <string name="settings__spelling__title" comment="Title of the Spelling screen">Rechtschreibprüfung</string>
-    <string name="settings__spelling__use_floris_config" comment="String used for the stub subtype in the system spell checker UI">Verwende Tastatur-Rechtschreibprüfung-Konfiguration</string>
-    <string name="pref__spelling__active_spellchecker__label">Inline-Textfeld-Rechtschreibprüfung</string>
     <string name="pref__spelling__active_spellchecker__summary_disabled">Systemweit deaktiviert. Es werden keine roten Linien in Textfeldern für falsche Wörter angezeigt. Zum Ändern antippen.</string>
     <string name="pref__spelling__active_spellchecker__summary_none">Keinen Inline-Rechtschreibprüfungs-Service eingestellt. Zum Ändern antippen.</string>
-    <string name="pref__spelling__active_spellchecker__summary_use_sys_lang_set">Die Option \"Sprachen\" in der System-Rechtschreibprüfungsansicht ist auf \"Systemsprachen verwenden\" eingestellt. Aufgrund eines Fehlers kann dies dazu führen, dass {app_name} keine Inline-Rechtschreibprüfung durchführt. Du kannst die Option zu \"{use_floris_config}\" ändern.</string>
-    <string name="pref__spelling__group_spellchecker_config__title" comment="Title of Fine-adjustment sub-group">Konfiguration</string>
     <string name="pref__spelling__language_mode__label" comment="Label of Language mode pref">Sprachen</string>
     <string name="pref__spelling__use_contacts__label" comment="Label of Use contact list pref">[NICHT IMPLEMENTIERT] Kontaktnamen benutzen</string>
     <string name="pref__spelling__use_contacts__summary" comment="Summary of Use contact list pref">Der Rechtschreibprüfer wird Namen aus dem Android Adressbuch einbinden</string>
